+++ src/gui/data_structures/static_mesh.cpp
@@ -1,9 +1,9 @@
 #include "static_mesh.hpp"
 
-#include "../gui/meshloader.hpp"
-#include "../logging.hpp"
-#include "../types.hpp"
-#include "mesh.hpp"
+#include "../meshloader.hpp"
+#include "../../logging.hpp"
+#include "../../types.hpp"
+#include "../../entity/mesh.hpp"
 
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
