+++ src/gui/main_gui.cpp
@@ -1,19 +1,19 @@
 #include "main_gui.hpp"
 
 #include "../entity/mesh.hpp"
-#include "../entity/static_mesh.hpp"
-#include "../entity/terrain_mesh.hpp"
 #include "../logging.hpp"
 #include "../util/files.hpp"
 #include "../world.hpp"
 #include "controls.hpp"
-#include "render/renderer.hpp"
-#include "shader.hpp"
 #include "data_structures/screen_data.hpp"
 #include "data_structures/sky_data.hpp"
+#include "data_structures/static_mesh.hpp"
+#include "data_structures/terrain_mesh.hpp"
+#include "gui_logging.hpp"
+#include "render/renderer.hpp"
 #include "render/shadow_map.hpp"
 #include "render/sky.hpp"
-#include "gui_logging.hpp"
+#include "shader.hpp"
 
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
