+++ src/gui/render/renderer.cpp
@@ -1,6 +1,6 @@
 #include "renderer.hpp"
 
-#include "../../entity/terrain_mesh.hpp"
+#include "../data_structures/terrain_mesh.hpp"
 #include "../../util/files.hpp"
 #include "../controls.hpp"
 #include "../meshloader.hpp"
@@ -47,12 +47,12 @@ MainRenderer::~MainRenderer() {
 }
 
 void
-MainRenderer::add_mesh(std::shared_ptr<MeshLoader::SingleComplexMesh> mesh) {
+MainRenderer::add_mesh(std::shared_ptr<MeshData::SingleComplexMesh> mesh) {
     singles_meshes_.push_back(std::move(mesh));
 }
 
 void
-MainRenderer::add_mesh(std::shared_ptr<MeshLoader::MultiComplexMesh> mesh) {
+MainRenderer::add_mesh(std::shared_ptr<MeshData::MultiComplexMesh> mesh) {
     multis_meshes_.push_back(std::move(mesh));
 }
 
@@ -131,7 +131,7 @@ MainRenderer::render(GLFWwindow* window) const {
     glBindTexture(GL_TEXTURE_2D, depth_texture_);
     glUniform1i(shadow_map_ID_, 1);
 
-    for (std::shared_ptr<MeshLoader::SingleComplexMesh> mesh : singles_meshes_) {
+    for (std::shared_ptr<MeshData::SingleComplexMesh> mesh : singles_meshes_) {
         glActiveTexture(GL_TEXTURE2);
         glBindTexture(GL_TEXTURE_1D, mesh->get_color_texture());
         glUniform1i(color_map_ID_, 2);
@@ -199,7 +199,7 @@ MainRenderer::render(GLFWwindow* window) const {
     glBindTexture(GL_TEXTURE_2D, depth_texture_);
     glUniform1i(shadow_map_ID_multi_, 1);
 
-    for (std::shared_ptr<MeshLoader::MultiComplexMesh> mesh : multis_meshes_) {
+    for (std::shared_ptr<MeshData::MultiComplexMesh> mesh : multis_meshes_) {
         glActiveTexture(GL_TEXTURE2);
         glBindTexture(GL_TEXTURE_1D, mesh->get_color_texture());
         glUniform1i(color_map_ID_multi_, 2);
